% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_variables.r
\name{calculate.stage.variables}
\alias{calculate.stage.variables}
\title{Calculating the stage variables}
\usage{
calculate.stage.variables(
  day,
  daycount,
  population,
  resident,
  immigrate.mean,
  immigrate.sd,
  emigrate.mean,
  emigrate.sd
)
}
\arguments{
\item{day}{int vector, days of observation, should be in ascending
order.}

\item{daycount}{int vector, counts of birds observed for each day,
should be coordinated with day vector.}

\item{population}{int or range vector of \code{c(min, max)} for
prediction, the total number of all the birds, including
migrants and residents.}

\item{resident}{numeric or range vector of \code{c(min, max)} for
prediction, the number of resident birds.}

\item{immigrate.mean}{numeric or range vector of \code{c(min,
max)} for prediction, the mean day of immigration.}

\item{immigrate.sd}{numeric or range vector of \code{c(min, max)}
for prediction, the standard diviation of immigration days.}

\item{emigrate.mean}{numeric or range vector of \code{c(min, max)}
for prediction, the mean day of emigration.}

\item{emigrate.sd}{numeric or range vector of \code{c(min, max)}
for prediciton, the standard diviation of emigration days.}
}
\value{
\code{calculate.stage.variables(day, daycount, population,
    immigrate.mean, immigrate.sd, emigrate.mean, emigrate.sd)}
    returns a list of output of calculated variables for the stage
    of the data inputed.
}
\description{
A wrapper function of \code{optim()} for the easier
    calculation of stage variables using the output function from
    \code{factory.stage.variables()}.
}
\examples{
data <- data.frame(
    day=c(11, 16, 20, 22, 24, 34, 39, 40, 47, 60),
    daycount=c(10, 65, 60, 45, 66, 52, 35, 50, 27, 0)
)

result <- calculate.stage.variables(
    day=data$day,
    daycount=data$daycount,
    population=85,
    resident = 0,
    immigrate.mean=c(10, 20),
    emigrate.mean=c(20, 50),
    immigrate.sd=c(4, 10),
    emigrate.sd=c(10, 30)
)

}
\seealso{
\code{factory.stage.variables()} for a lower level function to calculate,
    \code{optim()} for calculation using the output function of
    \code{factory.stage.variables()}.
}
