% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esetProstate}
\alias{esetProstate}
\title{Transcript-level abundance estimates in 14 Chinese prostate cancer patients}
\format{An \code{\link{ExpressionSet}} object.}
\source{
\url{http://pachterlab.github.io/lair/}
}
\usage{
esetProstate
}
\description{
A dataset containing 14 matched samples of tumoral prostate cancer and normal tissue, both derived from the same Chinese patient. The dataset has been prefiltered to reduce the computational burden of the vignette.
}
\keyword{datasets}

