% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_variables.r
\name{factory.stage.variables}
\alias{factory.stage.variables}
\title{Calculating of stage variables using \code{optim()}}
\usage{
factory.stage.variables(
  day,
  daycount,
  population,
  resident,
  immigrate.mean,
  immigrate.sd,
  emigrate.mean,
  emigrate.sd
)
}
\arguments{
\item{day}{int vector, days of observation, should be in ascending
order.}

\item{daycount}{int vector, counts of birds observed for each day,
should be coordinated with day vector.}

\item{population}{int or unset for prediction, the total number of
all the birds, including migrants and residents.}

\item{resident}{numeric or unset for prediction, the number of
resident birds.}

\item{immigrate.mean}{numeric or unset for prediction, the mean
day of immigration.}

\item{immigrate.sd}{numeric or unset for prediction, the standard
diviation of immigration days.}

\item{emigrate.mean}{numeric or unset for prediction, the mean day
of emigration.}

\item{emigrate.sd}{numeric or unset for prediciton, the standard
diviation of emigration days.}
}
\value{
\code{factory.stage.variables(day, daycount, population,
    immigrate.mean, immigrate.sd, emigrate.mean, emigrate.sd)}
    returns the function used for prediction with \code{optim}.
}
\description{
\code{factory.stage.variables} is a factory function
    that returns a function used for the calculation of variables
    for stage. The output function should be applied to
    \code{optim()} funtion.
}
\details{
The \code{factory.stage.variables} should be assigned
    with 'day' and 'daycount' parameters. Other parameters are
    optional. You should only assign the known parameters to the
    \code{factory.stage.variables} function. And other unknown
    parameters should be calculated with \code{optim()} function,
    with min and max range provided.
}
\examples{
data <- data.frame(
    day=c(11, 16, 20, 22, 24, 34, 39, 40, 47, 60),
    daycount=c(10, 65, 60, 45, 66, 52, 35, 50, 27, 0)
)

aimfunc <- factory.stage.variables(
    day=data$day, daycount=data$daycount, population=85
)

result <- optim(
    par=c(
        'immigrate.mean'=15, 'emigrate.mean'=40,
        'immigrate.sd'=5, 'emigrate.sd'=15),
    fn=aimfunc,
    method="L-BFGS-B",
    lower=c(
        'immigrate.mean'=10, 'emigrate.mean'=20,
        'immigrate.sd'=4, 'emigrate.sd'=10
    ),
    upper=c(
        'immigrate.mean'=25, 'emigrate.mean'=50,
        'immigrate.sd'=10, 'emigrate.sd'=30
    )
)

}
\seealso{
\code{optim()} for calculation using the output function of
    \code{factory.stage.variables()}, \code{calculate.stage.variables()}
    for easier calculation of stage variables.
}
