% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{getPConfirmation}
\alias{getPConfirmation}
\title{Return unadjusted confirmation hypothesis p-values from a \code{\link{stageRClass}} object.}
\usage{
\S4method{getPConfirmation}{stageR}(object)
}
\arguments{
\item{object}{an object of the \code{\link{stageRClass}} class.}
}
\description{
Return unadjusted confirmation hypothesis p-values from a \code{\link{stageRClass}} object.
}
\examples{
stageRObj <- stageR(pScreen=seq(1e-4,.5,length.out=1000), pConfirmation=matrix(runif(3000),nrow=1000,ncol=3))
getPConfirmation(stageRObj)
}

