% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hammer.eset}
\alias{hammer.eset}
\title{Hammer dataset}
\format{An \code{ExpressionSet} object.}
\source{
\url{http://bowtie-bio.sourceforge.net/recount/}
}
\usage{
hammer.eset
}
\description{
A gene expression dataset from an experiment on spinal nerve ligation in rats, comparing this treatment to control samples in two timepoints, i.e. two weeks and two months post treatment. 2 Biological replicates available in every treatment x time combination.
}
\references{
Hammer P, Banck MS, Amberg R, et al. mRNA-seq with agnostic splice site discovery for nervous system transcriptomics tested in chronic pain. Genome Research. 2010;20(6):847-860. doi:10.1101/gr.101204.109.
}
\keyword{datasets}
