% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate.r
\name{migrate.proportion}
\alias{migrate.proportion}
\alias{migrate.number}
\title{Calculating of stage parameters.}
\usage{
migrate.proportion(day, migrate.mean, migrate.sd)

migrate.number(day, population, migrate.mean, migrate.sd)
}
\arguments{
\item{day}{int, the day from the start of migration.}

\item{migrate.mean}{numeric, the mean day of immigration day or
emigration.}

\item{migrate.sd}{numeric, the standard diviation of immigration
days or emigration days.}

\item{population}{int, the total number of all the migrated birds.}
}
\value{
\code{migrate.proportion(day, migrate.mean, migrate.sd)}
    returns the proportion of the migrated birds (immigration or
    emigration).

\code{migrate.number(day, population, migrate.mean,
    migrate.sd)} returns the number of the migrated birds
    (immigration or emigration).
}
\description{
\code{migrate.proportion()} calculates the
    immigration or emmigration proportion of whole population
    using normal distribution.

\code{migrate.number()} calculates the immigration or
    emmigration numbers. \code{migrate.number()} returns the
    proportion result from \code{migrate.proportion()} times the
    whole population number.
}
\examples{
migrate.proportion(10, 15, 8)

migrate.number(10, 5000, 15, 8)

}
\seealso{
[calculate.stage.variables()] for the estimation of
    important stage variables.
}
