% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandpiper.r
\docType{data}
\name{sandpiper}
\alias{sandpiper}
\title{Staging data of sandpiper.}
\format{
A data frame with 77 rows and 2 columns.
\describe{
  \item{day}{the integer vector of days.}
  \item{daycount}{the integer vector of the sandpiper counts for each day.}
}
}
\source{
Thompson, J.J. (1993). Modeling the Local Abundance of Shorebirds Staging on Migration. Theoretical Population Biology 44, 299–315.
}
\usage{
sandpiper
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
