% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage.r
\name{stage.number}
\alias{stage.number}
\title{Calulating of staging bird numbers by stage variables.}
\usage{
stage.number(
  day,
  population,
  resident,
  immigrate.mean,
  immigrate.sd,
  emigrate.mean,
  emigrate.sd
)
}
\arguments{
\item{day}{int, the day from the start of migration.}

\item{population}{int, the total number of all the birds,
including migrants and residents.}

\item{resident}{numeric, the number of
resident birds.}

\item{immigrate.mean}{numeric, the mean day of immigration.}

\item{immigrate.sd}{numeric, the standard diviation of immigration
days.}

\item{emigrate.mean}{numeric, the mean day of emigration.}

\item{emigrate.sd}{numeric, the standard diviation of emigration
days.}
}
\value{
\code{stage.number(day, population, immigrate.mean,
    immigrate.sd, emigrate.mean, emigrate.sd)} returns the number
    of the staging birds of gaven days.
}
\description{
Function for calculation of staging bird numbers
    using normal distributions of immigration and emigration.
}
\examples{
stage.number(
    day=10, population=5000,
    immigrate.mean=15, immigrate.sd=8,
    emigrate.mean=45, emigrate.sd=20
)

}
