% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_variables.r
\name{factory.stage.variables}
\alias{factory.stage.variables}
\alias{calculate.stage.variables}
\title{Calculating of stage variables using \code{optim()}}
\usage{
factory.stage.variables(
  day,
  daycount,
  population,
  immigrate.mean,
  emigrate.mean,
  immigrate.sd,
  emigrate.sd
)

calculate.stage.variables(
  day,
  daycount,
  population,
  immigrate.mean,
  emigrate.mean,
  immigrate.sd,
  emigrate.sd
)
}
\arguments{
\item{day}{int vector, days of observation, should be in ascending
order.}

\item{daycount}{int vector, counts of birds observed for each day,
should be coordinated with day vector.}

\item{population}{int or range vector of \code{c(min, max)} for
prediction, the total number of all the migrated birds.}

\item{immigrate.mean}{numeric or range vector of \code{c(min,
max)} for prediction, the mean day of immigration.}

\item{emigrate.mean}{numeric or range vector of \code{c(min, max)}
for prediction, the mean day of emigration.}

\item{immigrate.sd}{numeric or range vector of \code{c(min, max)}
for prediction, the standard diviation of immigration days.}

\item{emigrate.sd}{numeric or range vector of \code{c(min, max)}
for prediciton, the standard diviation of emigration days.}
}
\value{
\code{factory.stage.variables(day, daycount, population,
    immigrate.mean, immigrate.sd, emigrate.mean, emigrate.sd)}
    returns the function used for prediction with \code{optim}.

\code{calculate.stage.variables(day, daycount, population,
    immigrate.mean, immigrate.sd, emigrate.mean, emigrate.sd)}
    returns a list of output of calculated variables for the stage
    of the data inputed.
}
\description{
\code{factory.stage.variables} is a factory function
    that returns a function used for the calculation of variables
    for stage. The output function should be applied to
    \code{optim()} funtion.

A wrapper function of \code{optim()} for the easier
    calculation of stage variables using the output function from
    \code{factory.stage.variables()}.
}
\details{
The \code{factory.stage.variables} should be assigned
    with 'day' and 'daycount' parameters. Other parameters are
    optional. You should only assign the known parameters to the
    \code{factory.stage.variables} function. And other unknown
    parameters should be calculated with \code{optim()} function,
    with min and max range provided.
}
\examples{
data <- data.frame(
    day=c(11, 16, 20, 22, 24, 34, 39, 40, 47, 60),
    daycount=c(10, 65, 60, 45, 66, 52, 35, 50, 27, 0)
)

aimfunc <- factory.stage.variables(
    day=data$day, daycount=data$daycount, population=85
)

result <- optim(
    par=c(
        'immigrate.mean'=15, 'emigrate.mean'=40,
        'immigrate.sd'=5, 'emigrate.sd'=15),
    fn=aimfunc,
    method="L-BFGS-B",
    lower=c(
        'immigrate.mean'=10, 'emigrate.mean'=20,
        'immigrate.sd'=4, 'emigrate.sd'=10
    ),
    upper=c(
        'immigrate.mean'=25, 'emigrate.mean'=50,
        'immigrate.sd'=10, 'emigrate.sd'=30
    )
)

data <- data.frame(
    day=c(11, 16, 20, 22, 24, 34, 39, 40, 47, 60),
    daycount=c(10, 65, 60, 45, 66, 52, 35, 50, 27, 0)
)

result <- calculate.stage.variables(
    day=data$day,
    daycount=data$daycount,
    population=85,
    immigrate.mean=c(10, 20),
    emigrate.mean=c(20, 50),
    immigrate.sd=c(4, 10),
    emigrate.sd=c(10, 30)
)

}
\seealso{
[optim()] for calculation using the output function of
    \code{factory.stage.variables}.
}
