% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{adjustedAlphaLevel,stageR-method}
\alias{adjustedAlphaLevel,stageR-method}
\alias{adjustedAlphaLevel,stageRTx-method}
\title{Get adjusted significance level from the screening stage.}
\usage{
\S4method{adjustedAlphaLevel}{stageR}(object)

\S4method{adjustedAlphaLevel}{stageRTx}(object)
}
\arguments{
\item{object}{an object of the \code{\link{stageRClass}} class.}
}
\value{
Scalar, the adjusted significance level from the screening stage.
}
\description{
This functions returns the adjusted significance level from the screening stage that should be used to compare confirmation stage FWER adjusted p-values against.
}
\details{
The adjusted significance level is calculated as the fraction of significant features in the screening stage multiplied the alpha level.
}
\section{Methods (by class)}{
\itemize{
\item \code{stageR}: Get adjusted significance level from the screening stage.

\item \code{stageRTx}: Get adjusted significance level from the screening stage.
}}

\examples{
pScreen=c(seq(1e-10,1e-2,length.out=100),seq(1e-2,.2,length.out=100),seq(.2,1,length.out=100))
names(pScreen)=paste0("gene",1:300)
pConfirmation=matrix(runif(900),nrow=300,ncol=3)
dimnames(pConfirmation)=list(paste0("gene",1:300),c("H1","H2","H3"))
stageRObj <- stageR(pScreen=pScreen, pConfirmation=pConfirmation, pScreenAdjusted=FALSE)
stageRObj <- stageWiseAdjustment(stageRObj, method="holm", alpha=0.05)
adjustedAlphaLevel(stageRObj)
# @method stageR-method
}
\references{
Van den Berge K., Soneson C., Robinson M.D., Clement L. 2017. A general and powerful stage-wise testing procedure for differential expression and differential transcript usage. \url{http://biorxiv.org/content/early/2017/02/16/109082}

R. Heller, E. Manduchi, G. R. Grant, and W. J. Ewens, "A flexible two-stage procedure for identifying gene sets that are differentially expressed." Bioinformatics (Oxford, England), vol. 25, pp. 1019-25, 2009.
}
\seealso{
\code{\link{stageR}}, \code{\link{stageRClass}}
# @name adjustedAlphaLevel
# @rdname adjustedAlphaLevel
}
