% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{adjustedAlphaLevel}
\alias{adjustedAlphaLevel}
\title{Get adjusted significance level from the screening stage.}
\usage{
\S4method{adjustedAlphaLevel}{stageR}(object)
}
\arguments{
\item{object}{an object of the \code{\link{stageRClass}} class.}
}
\description{
This functions returns the adjusted significance level from the screening stage that should be used to compare confirmation stage FWER adjusted p-values to.
}
\details{
The adjusted significance level is calculated as the fraction of significant features in the screening stage times the alpha level.
}
\examples{
pScreen=c(seq(1e-10,1e-2,length.out=100),seq(1e-2,.2,length.out=100),seq(.2,1,length.out=100))
stageRObj <- stageR(pScreen=pScreen, pConfirmation=matrix(runif(3000),nrow=1000,ncol=3))
adjustedP <- stageWiseAdjustment(stageRObj, method="holm", alpha=0.05)
adjustedAlphaLevel(adjustedP)
}

