% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{getAdjustedPValues}
\alias{getAdjustedPValues}
\title{Retrieve the stage-wise adjusted p-values.}
\usage{
\S4method{getAdjustedPValues}{stageR}(object, ...)
}
\arguments{
\item{object}{an object of the \code{\link{stageRClass}} class.}
}
\description{
This functions returns the stage-wise adjusted p-values for an object from the  \code{\link{stageRClass}} class. Note, that the p-values should have been adjusted with the \code{\link{stageWiseAdjustment}} function prior to calling this function.
}
\details{
The function returns FDR adjusted p-values for the screening hypothesis and stage-wise adjusted p-values for the confirmation hypothesis p-values. For features that were not significant in the screening hypothesis, the confirmation stage adjusted p-values are set to 1.
}
\examples{
pScreen=c(seq(1e-10,1e-2,length.out=100),seq(1e-2,.2,length.out=100),seq(.2,1,length.out=100))
stageRObj <- buildStageR(pScreen=pScreen, pConfirmation=matrix(runif(3000),nrow=1000,ncol=3))
adjustedP <- stageWiseAdjustment(stageRObj, method="holm", alpha=0.05)
getAdjustedPValues(adjustedP)
}

