% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{getSignificantGenes}
\alias{getSignificantGenes}
\title{Return significant genes when performing transcript level analysis.}
\usage{
\S4method{getSignificantGenes}{stageR}(object)
}
\arguments{
\item{object}{an object of the \code{\link{stageRClass}} class.}
}
\description{
This functions returns a matrix with significant genes by aggregated testing of its respective transcripts.
}
\examples{
#make identifiers linking transcripts to genes
set.seed(1)
genes=paste0("gene",sample(1:200,1000,replace=TRUE))
nGenes=length(table(genes))
transcripts=paste0("tx",1:1000)
tx2gene=data.frame(transcripts,genes)
#gene-wise q-values
pScreen=c(seq(1e-10,1e-2,length.out=nGenes-100),seq(1e-2,.2,length.out=50),seq(50))
names(pScreen)=names(table(genes)) #discards genes that are not simulated
pConfirmation=matrix(runif(1000),nrow=1000,ncol=1)
rownames(pConfirmation)=transcripts
stageRObj <- buildStageR(pScreen=pScreen, pConfirmation=pConfirmation ,pScreenAdjusted=TRUE)
adjustedP <- stageWiseAdjustment(stageRObj, method="dte", alpha=0.05, tx2gene=tx2gene)
head(getSignificantGenes(adjustedP))
}

