% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{stageR}
\alias{stageR}
\title{Create stageR object}
\usage{
stageR(pScreen, pConfirmation, pScreenAdjusted = FALSE)
}
\arguments{
\item{pScreen}{A vector of screening hypothesis p-values.}

\item{pConfirmation}{A matrix of confirmation hypothesis p-values. The number of rows should be equal to the length of \code{pScreen}.}

\item{pScreenAdjusted}{logical, indicating whether the supplied p-values for the screening hypothesis have already been adjusted for multiplicity according to the FDR.}
}
\value{
An instance of an object of the \code{\link{stageRClass}}
}
\description{
Constructor function for \code{\link{stageRClass}}. A stageR class is a class used for stage-wise analysis in high throughput settings.
In its most basic form, it consists of a vector of p-values for the screening hypothesis and a matrix of p-values for the confirmation hypotheses.
}
\examples{
# create a \\code{\\link{stageRClass}} object
stageRObj <- stageR(pScreen=runif(10), pConfirmation=matrix(runif(30),nrow=10,ncol=3))
}
\references{
Van den Berge K., Soneson C., Robinson M.D., Clement L.
}

