% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{stageR}
\alias{stageR}
\alias{stageRTx}
\title{Create stageR object}
\usage{
stageR(pScreen, pConfirmation, pScreenAdjusted = FALSE)

stageRTx(pScreen, pConfirmation, pScreenAdjusted = FALSE, tx2gene)
}
\arguments{
\item{pScreen}{A vector of screening hypothesis p-values.}

\item{pConfirmation}{A matrix of confirmation hypothesis p-values. The number of rows should be equal to the length of \code{pScreen}.}

\item{pScreenAdjusted}{logical, indicating whether the supplied p-values for the screening hypothesis have already been adjusted for multiplicity according to the FDR.}

\item{tx2gene}{Only applicable for transcript-level analysis. A \code{\link[base]{data.frame}} with transcript IDs in the first columns and gene IDs in the second column. The rownames from \code{pConfirmation} must be contained in the transcript IDs from \code{tx2gene}, and the names from \code{pScreen} must be contained in the gene IDs.}
}
\value{
An instance of an object of the \code{\link{stageRClass}}
}
\description{
Constructor function for \code{\link{stageRClass}}. A stageR class is a class used for stage-wise analysis in high throughput settings.
In its most basic form, it consists of a vector of p-values for the screening hypothesis and a matrix of p-values for the confirmation hypotheses.
}
\examples{
# create a \\code{\\link{stageRClass}} object
stageRObj <- stageR(pScreen=runif(10), pConfirmation=matrix(runif(30),nrow=10,ncol=3))
stageRObj <- stageRTx(pScreen=runif(10), pConfirmation=matrix(runif(30),nrow=10,ncol=3), tx2gene=data.frame(transcripts=paste0("transcript",1:10),genes=paste0("gene",rep(1:2,each=5))))
}
\references{
Van den Berge K., Soneson C., Robinson M.D., Clement L.
}

